Update code to be more modular
